fix startup error by waiting for app ready
